+++ src/common/create_component/test.py
@@ -1,7 +1,7 @@
 import os
 import subprocess
 from os import path
-from ruamel.yaml import YAML
+import ruamel.yaml as yaml
 
 ## VIASH START
 meta = {
@@ -10,7 +10,7 @@ meta = {
 ## VIASH END
 
 opv2 = f"{meta['resources_dir']}/openproblems-v2"
-output_path = f"{opv2}/src/label_projection/methods/test_method"
+output_path = f"{opv2}/src/tasks/label_projection/methods/test_method"
 
 cmd = [
     meta['executable'],
@@ -21,7 +21,14 @@ cmd = [
 ]
 
 print('>> Running the script as test', flush=True)
-out = subprocess.run(cmd, check=True, cwd=opv2)
+out = subprocess.run(cmd, stderr=subprocess.STDOUT, cwd=opv2)
+
+if out.stdout:
+    print(out.stdout)
+
+if out.returncode:
+    print(f"script: '{cmd}' exited with an error.")
+    exit(out.returncode)
 
 print('>> Checking whether output files exist', flush=True)
 assert os.path.exists(output_path), "Output dir does not exist"
@@ -33,12 +40,11 @@ script_f = path.join(output_path, "script.py")
 assert os.path.exists(script_f), "Script file does not exist"
 
 print('>> Checking file contents', flush=True)
-yaml = YAML()
 with open(conf_f) as f:
-    conf_data = yaml.load(f)
+    conf_data = yaml.safe_load(f)
 
 assert conf_data['functionality']['name'] == 'test_method', "Name should be equal to 'test_method'"
-assert conf_data['platforms'][0]['image'] == 'python:3.10', "Python image should be equal to python:3.10"
+# assert conf_data['platforms'][0]['image'] == 'python:3.10', "Python image should be equal to python:3.10"
 
 
 print('All checks succeeded!', flush=True)